pyproject.toml
[tool.poetry]
name = "ocr-test-app"
version = "0.1.0"
description = "App per testare e confrontare diversi OCR con AI"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
easyocr = "^1.7.0"
paddleocr = "^2.7.3"
pytesseract = "^0.3.10"
Pillow = "^10.0.0"
opencv-python = "^4.8.0"
streamlit = "^1.28.0"
matplotlib = "^3.7.2"
numpy = "^1.24.3"
pandas = "^2.0.3"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
httpx = "^0.27.0"
aiofiles = "^23.2.0"
plotly = "^5.17.0"
gradio = "^4.44.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"